feat: Migrate the ADMIN system from ADMIN_KEY to an ADMIN model authority system
BREAKING CHANGE: The previous ADMIN system, ADMIN_KEY, was a TOKEN generated by reset.sh. Any request made with that TOKEN would be deemed as a request made by an ADMIN.
Now, with the new ADMIN model authority system, the ADMIN is an user, therefore it has an username (default: "admin") and password (generated by reset.sh").
IMPORTANT: When a new user is created with the "President" role, it is set to be the new ADMIN (the last ADMIN credentials are overwritten).